requirements.txt
numpy==1.18.1
opencv-contrib-python==4.1.2.30
Pillow==8.0.1
pytesseract==0.3.1
